Attempt to port some pod-generating code from previous work.  Problematic.
This is code that more or less verbatim worked... when linked against the kubernetes source tree directly, and pinned to a version several months ago.

Since then, it seems that the ObjectMeta type has been moved to this new package... in a new library... that despite clearly being in my source tree go, won't let me link directly since it's in a vendor dir, so symlinks to the rescue, but thereafter it wants to pull a world down on my head of other libraries... I could keep symlinking for days, but...

What we're actually going to do is a barrel roll here, and simply try not to reference that package explicitly, instead traversing members to get to things like that.  This feels fairly ridiculous.  But preliminarily, it seems to be possible; and generically, it seems that the less we know and touch k8s code, the better for maintainability on our side, due to their spectacular churn rate and lack of normalized types in source tree.

Signed-off-by: Eric Myhre <hash@exultant.us>