Link sufficient libraries to be able to import the kubernetes yaml/json parsers.
We want to match their behavior for loading files as precisely as possible.

Libraries are being placed on the gopath via symlink to their location vendored inside the client-go repo.  This is because importing from another library's 'vendor' package is forbidden by the compiler.  However, it seems perfectly correct to reuse exactly those libraries: I certainly don't have any opinion on their versions except for precisely that they make the k8s opinions of them.  Nor is it reasonable to make more submodules for them just of a misplaced sense of purity.

Signed-off-by: Eric Myhre <hash@exultant.us>